broken work in progress state
all files are now loaded from the bogus filesystem, including the main
binary.

this works ok, but since it's longer using initrd, the first free address is
set to the start of where we're loading the kernel (whoops), instead of the
end of initrd like it used to be

the FreeBSD rtld wants to load programs to their preferred address, and
those tend to be pretty low (4MB for i386, 2MB for amd64), so I need to make
sure the kernel isn't too big; which means kicking the filesystem out again.